requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=1.0.0

# RAG and document processing
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
pypdf>=3.0.0

# AI Models
transformers>=4.30.0
torch>=2.0.0
google-generativeai>=0.3.0

# Evaluation
scikit-learn>=1.3.0
rouge-score>=0.1.2

# Data processing
openpyxl>=3.1.0